requirements.txt
aiogram==2.23.1
aiohttp==3.8.3
aiosignal==1.3.1
altgraph==0.17.2
amqp==5.1.1
APScheduler==3.9.1.post1
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
bottle==0.12.23
bottle-websocket==0.2.9
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.15
Eel==0.14.0
emoji==2.0.0
frozenlist==1.3.3
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
h11==0.13.0
idna==3.4
kombu==5.2.4
lxml==4.9.1
magic-filter==1.0.9
multidict==6.0.4
outcome==1.2.0
pefile==2022.5.30
prompt-toolkit==3.0.36
pycparser==2.21
pyinstaller==5.3
pyinstaller-hooks-contrib==2022.8
pyparsing==3.0.9
PySocks==1.7.1
python-dotenv==0.20.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pywin32-ctypes==0.2.0
rabbitmq==0.2.0
requests==2.28.1
selenium==4.4.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
whichcraft==0.6.1
wsproto==1.1.0
yarl==1.8.2
zope.event==4.5.0
zope.interface==5.4.0
